Add files via upload
  Business Problem:-
In recent years, City Hotel and Resort hotels have seemed high cancellation rates. Each Hotel is now dealing with a number of issues as a result, including fewer revenues and less than-ideal hotel room use. Consequently, lowering cancellation rates is both hotels’ primary goal in order to increase their efficiency in generating revenue, and for us to offer business advice to address this problem.
The analysis of hotel booking cancellations as well as other factors that have to bring on their business and yearly revenue generation are the main topics of this report.

   Assumptions:-
1.	No unusual occurrences between 2015 and 2017 will have a substantial impact on the data used.
2.	The information is still current and can be used to analyze a hotel’s possible plan in an efficient manner.
3.	There are no unanticipated negatives to the hotel employing any advised technique.
4.	The hotels are not currently using any of the suggested solutions.
5.	The biggest factor affecting the effectiveness of earning income is booking cancellations.
6.	Cancellation results in vacant rooms for the booked length of time.
7.	Clients make hotel reservations the same year they make cancellations.

  Research Question:-
1.	What are the variables that affect hotel reservation cancellations?
2.	How can we make hotel reservation cancellations better?
3.	How will hotels be assisted in making pricing and promotion decisions?

  Hypothesis:-
1.	More cancellations occur when prices are higher.
2.	When there is a longer waiting list, customers tend to cancel more frequently.
3.	The majority of clients are coming from offline travel agents to make their reservations.

                 
  The accompanying bar graph the percentage of reservations that are canceled And those that are not. It is obvious that there are still a significant number of reservations that have not been canceled. There are still 37% of clients who canceled their reservations, which has a significant impact on the hotel’s earnings.
              
In comparison to resort hotels, city hotels have bookings. It’s possible that resort hotels are more expensive than those in cities.

              
The line graph above shows that, on certain days, the average daily rate for a city hotel Is less than that of a resort hotel, and on other days, it is even less. It goes without saying that weekends may see a rise in resort hotel rates.
        

We have developed the grouped bar graph to analyze the months with the Highest and lowest reservation levels according to reservation status. As can be seen, both the number of confirmed reservations and the number of canceled Reservations are the largest in the month of August. Whereas January is the month with the most canceled reservations.


           
       This bar graph demonstrates that cancellations are most common when prices are greatest and least common when they are lowest. Therefore, the cost of the accommodation is solely responsible for the cancellation.
Now, let’s see which country has the highest reservation canceled. The top country is Portugal with the highest number of cancellations.
 
Let’s check the area from where guests are visiting the hotels and making reservations. Is it coming from direct or groups? Online or Offline Travel Agents? Around 46%   of the clients come from online travel agencies. Whereas 27% come from groups. Only 4% of clients book hotels directly by visiting them and making reservations.
 
As seen in the graph, reservations are canceled when the average daily rate is higher than when it is not canceled. It clearly proves all the above analysis that a higher price leads to higher cancellations.

Suggestion:-
1.	Cancellation rates as the price does. In order to prevent cancellation of Reservations, hotels could work on their pricing strategies and try to lower the Rates for specific hotels based on locations. They can also provide the same Discounts to the customers.
2.	     As the ratio of cancellations and not cancellations of the resort hotel is higher in the resort hotel than in the city hotels. So the hotels should provide a reasonable discount on the room price on Weekends or on holidays.
3.	In the month of January, hotels can start campaigns or marketing with a Reasonable amount to increase their revenue as cancellation is the highest in this month.
4.	They can also increase the quality of their hotels and their services mainly In Portugal to reduce the cancellation rate.